build: autocompilation avec choix des révisions et configuration de la compilation

+++ menu.sh
@@ -32,6 +32,7 @@ echo "7   Tuer tout les processus"
 echo "8   Nettoyer les sources"
 echo "9   Voir le disque en hexadécimal"
 echo "c   Changer la version de developpement"
+echo "o   Changer les options de développement"
 echo "0   Quitter"
 echo "------------------------------------------"
 echo "Choisissez l'action à réaliser..."
@@ -48,25 +49,44 @@ case "${answer}" in
 7) ./make.sh killer;;
 8) ./make.sh clean;;
 9) ./make.sh view|more;;
-c) echo "Version disponibles:"
+o*) ./make.sh config
+./make.sh tool
+;; 
+c*) echo "Version disponibles:"
 SELECT=$(git branch|grep "*"|tr -d "* ")
 NUM=1
+echo "XX  Hash d'une révision particulière" 
 while read LINE
 do
    echo "${NUM}   ${LINE}"
    (( NUM++ )) 
 done < <(git branch -lr|tr -d "* "|grep -v HEAD|sed s/".*origin\/"//)
-read ANSWER
-CHOOSE=$(git branch -lr|tr -d "* "|grep -v HEAD|sed s/".*origin\/"//|tr "\n" " "|cut -d " " -f${ANSWER})
-echo "vous avez selectionné $ALL..."
-git checkout $CHOOSE
-git clean -fd
-git reset --hard
+read -p"?" ANSWER
+COUNT=$(echo -n ${ANSWER}|wc -c)
+if [ "${COUNT}" == "1" ]; then
+	CHOOSE=$(git branch -lr|tr -d "* "|grep -v HEAD|sed s/".*origin\/"//|tr "\n" " "|cut -d " " -f${ANSWER})
+	echo "vous avez selectionné une branche $CHOOSE..."
+else
+	EXIST=$(git show ${ANSWER})
+	if [ "${EXIST}" != "" ]; then
+		CHOOSE=${ANSWER}
+		echo "vous avez selectionné une révision ${choose}..."
+	fi
+fi
+if [ "${CHOOSE}" != "" ]; then
+	echo "*** Application de la version ${CHOOSE}"
+	git checkout $CHOOSE
+	git clean -fd
+	git reset --hard
+	git pull -f
+	./make.sh tool
+fi
 read
 ;;
 esac
 
 
 done
+
